requirements.txt
Django>=4.1.4
psycopg2>=2.9.5
pytablewriter>=0.64.2
requests>=2.28.1
ckanapi>=4.7
pillow>=9.3.0
json2html>=1.3.0
python-dateutil>=2.8.2
beautifulsoup4
celery
